// lib/types.ts
export type BadgeType = 'TopAnalyst' | 'Verified' | 'EarlyBird' | 'Whale' | 'Influencer' | 'RisingStar';

export interface Usuario {
  id: string;
  nombre: string;
  usuario: string;
  email?: string;
  password?: string;
  avatar: string;
  avatarUrl?: string;
  banner?: string;
  esPro: boolean;
  esVerificado?: boolean;
  rol: 'admin' | 'user' | 'ceo' | 'programador' | 'diseñador' | 'colaborador' | 'trader_inicial' | 'trader_experimentado' | 'cursante' | 'vip' | 'estudiante' | 'visitante';
  role: number;
  xp: number;
  level: number;
  biografia?: string;
  instagram?: string;
  seguidores: string[];
  siguiendo: string[];
  aportes: number;
  accuracy: number;
  reputation: number;
  badges: BadgeType[];
  watchlist: string[];
  estadisticas: {
    tasaVictoria: number;
    factorBeneficio: number;
    pnlTotal: number;
    layoutOrder?: string[];
  };
  saldo: number;
  fechaRegistro?: string;
  diasActivos?: number;
  ultimoLogin?: string;
  watchedClasses?: string[];
  mascotas?: any[];
  medallas?: any[];
  medianas?: any[];
  progreso?: Record<string, boolean>;
  lastSeen?: number;
  isBlocked?: boolean;
  blockedIPs?: string[];
  verificationPending?: boolean;
  avatarFrame?: string;
  streakReward?: string;
  userNumber?: number;
}